+++ frontend/employee/src/views/form/fields/SimpleView.jsx
