+++ frontend/employee/src/views/form/fields/SliderView.jsx
